query/requirements.txt
arrow==0.17.0
beautifulsoup4==4.9.3
bibtexparser==1.2.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
fake-useragent==0.1.11
free-proxy==1.0.2
future==0.18.2
idna==2.10
loguru==0.5.3
lxml==4.6.1
notify-py==0.3.0
proxyscrape==0.3.0
pycparser==2.20
pyOpenSSL==19.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
PyYAML==5.3.1
requests==2.24.0
scholarly==0.5.1
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
stem==1.8.0
urllib3==1.25.11
